// packages/spec/src/generated/envelope.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SignedEnvelope {
  manifest: {
    [k: string]: unknown;
  };
  signature: EnvelopeSignature;
}
export interface EnvelopeSignature {
  alg: "EdDSA";
  kid: string;
  sig: string;
}